Move EnqueuedDocument-type tests
